Split lib.rs into read.rs and write.rs
